requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
requests==2.28.2
tiktoken==0.2.0
beautifulsoup4==4.12.0
fake-useragent==1.1.3
msal==1.22.0